fix ata driver

+++ include/ata.h
@@ -5,20 +5,20 @@
 #include <port.h>
 #include <display.h>
 
-typedef struct {                // LBA28 | LBA48
-    uint8_t data_port;          // 16bit | 16 bits
-    uint8_t error_port;         // 8 bit | 16 bits
-    uint8_t sector_count_port;  // 8 bit | 16 bits
-    uint8_t lba_lo_port;        // 8 bit | 16 bits
-    uint8_t lba_mid_port;       // 8 bit | 16 bits
-    uint8_t lba_hi_port;        // 8 bit | 16 bits
-    uint8_t device_port;        // 8 bit
-    uint8_t command_port;       // 8 bit
-    uint8_t control_port;       // 8 bit
+typedef struct {                 // LBA28 | LBA48
+    uint32_t data_port;          // 16bit | 16 bits
+    uint32_t error_port;         // 8 bit | 16 bits
+    uint32_t sector_count_port;  // 8 bit | 16 bits
+    uint32_t lba_lo_port;        // 8 bit | 16 bits
+    uint32_t lba_mid_port;       // 8 bit | 16 bits
+    uint32_t lba_hi_port;        // 8 bit | 16 bits
+    uint32_t device_port;        // 8 bit
+    uint32_t command_port;       // 8 bit
+    uint32_t control_port;       // 8 bit
     char is_master;
 } ata_t;
 
-void init_ata(ata_t *ata, uint8_t port, char is_master);
+void init_ata(ata_t *ata, uint32_t port, char is_master);
 void indentify_ata_device(ata_t *ata);
 
 #endif
